+++ Sources/Scru128/Scru128Id.swift
@@ -1,17 +1,36 @@
 /// Represents a SCRU128 ID and provides converters and comparison operators.
 public struct Scru128Id: LosslessStringConvertible {
+  /// **Deprecated.** Use ``byteArray`` instead. From v3.0.0, this property will return a tuple of
+  /// 16 byte values instead of a byte array.
+  ///
   /// Returns a 16-byte byte array containing the 128-bit unsigned integer representation in the
   /// big-endian (network) byte order.
   public let bytes: [UInt8]
 
-  /// Creates an object from a byte array that represents a 128-bit unsigned integer.
+  /// Creates an object from a byte tuple that represents a 128-bit unsigned integer.
   ///
   /// - Parameter bytes: A 16-byte byte array that represents a 128-bit unsigned integer in the
   ///                    big-endian (network) byte order.
+  public init(
+    _ bytes: (
+      UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8,
+      UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8
+    )
+  ) {
+    self.bytes = [
+      bytes.0, bytes.1, bytes.2, bytes.3, bytes.4, bytes.5, bytes.6, bytes.7,
+      bytes.8, bytes.9, bytes.10, bytes.11, bytes.12, bytes.13, bytes.14, bytes.15,
+    ]
+  }
+
+  /// Creates an object from a byte array that represents a 128-bit unsigned integer.
+  ///
+  /// - Parameter byteArray: A 16-byte byte array that represents a 128-bit unsigned integer in the
+  ///                        big-endian (network) byte order.
   /// - Precondition: The byte length of the argument must be 16.
-  public init(_ bytes: [UInt8]) {
-    precondition(bytes.count == 16)
-    self.bytes = bytes
+  public init(_ byteArray: [UInt8]) {
+    precondition(byteArray.count == 16)
+    bytes = byteArray
   }
 
   /// Creates an object from field values.
@@ -50,10 +69,10 @@ public struct Scru128Id: LosslessStringConvertible {
 
   /// Creates an object from a 25-digit string representation.
   public init?(_ description: String) {
-    guard let bs = Self.parse(description) else {
+    guard let byteArray = Self.parse(description) else {
       return nil
     }
-    bytes = bs
+    bytes = byteArray
   }
 
   /// Builds the 16-byte big-endian byte array representation from a string.
@@ -76,10 +95,7 @@ public struct Scru128Id: LosslessStringConvertible {
       var minIndex = 99  // any number greater than size of output array
       for i in stride(from: -5, to: 25, by: 10) {
         // implement Base36 using 10-digit words
-        var carry: UInt64 = 0
-        for e in src[(i < 0 ? 0 : i)..<(i + 10)] {
-          carry = (carry * 36) + UInt64(e)
-        }
+        var carry: UInt64 = src[(i < 0 ? 0 : i)..<(i + 10)].reduce(0) { $0 * 36 + UInt64($1) }
 
         // iterate over output array from right to left while carry != 0 but at least up to place
         // already filled
@@ -98,6 +114,10 @@ public struct Scru128Id: LosslessStringConvertible {
       return dst
     }
   }
+
+  /// Returns a 16-byte byte array containing the 128-bit unsigned integer representation in the
+  /// big-endian (network) byte order.
+  public var byteArray: [UInt8] { bytes }
 
   /// Returns the 48-bit `timestamp` field value.
   public var timestamp: UInt64 { subUInt(0..<6) }
@@ -147,11 +167,7 @@ public struct Scru128Id: LosslessStringConvertible {
 
   /// Returns a part of `bytes` as an unsigned integer.
   private func subUInt<T: UnsignedInteger>(_ range: Range<Int>) -> T {
-    var buffer: T = 0
-    for e in bytes[range] {
-      buffer = (buffer << 8) | T(e)
-    }
-    return buffer
+    range.reduce(0) { $0 << 8 | T(bytes[$1]) }
   }
 }
 
@@ -179,14 +195,14 @@ private let decodeMap: [UInt8] = [
 ]
 
 extension Scru128Id: Comparable, Hashable {
-  public static func == (lhs: Scru128Id, rhs: Scru128Id) -> Bool {
-    return lhs.bytes == rhs.bytes
-  }
+  public static func == (lhs: Scru128Id, rhs: Scru128Id) -> Bool { lhs.bytes == rhs.bytes }
 
   public static func < (lhs: Scru128Id, rhs: Scru128Id) -> Bool {
-    for i in 0..<lhs.bytes.count {
-      if lhs.bytes[i] != rhs.bytes[i] {
-        return lhs.bytes[i] < rhs.bytes[i]
+    for i in 0..<16 {
+      let lft = lhs.bytes[i]
+      let rgt = rhs.bytes[i]
+      if lft != rgt {
+        return lft < rgt
       }
     }
     return false